Change sizeMp4 and sizeWebp params from long to long long

+++ src/megachatapi.cpp
@@ -1056,7 +1056,7 @@ int MegaChatApi::loadAttachments(MegaChatHandle chatid, int count)
     return pImpl->loadAttachments(chatid, count);
 }
 
-MegaChatMessage* megachat::MegaChatApi::sendGiphy(MegaChatHandle chatid, const char* srcMp4, const char* srcWebp, long sizeMp4, long sizeWebp, int width, int height, const char* title)
+MegaChatMessage* megachat::MegaChatApi::sendGiphy(MegaChatHandle chatid, const char* srcMp4, const char* srcWebp, long long sizeMp4, long long sizeWebp, int width, int height, const char* title)
 {
     return pImpl->sendGiphy(chatid, srcMp4, srcWebp, sizeMp4, sizeWebp, width, height, title);
 }
